Move Shared Services into Folder
