+++ Assets/Scripts/GamePlay/GamePlayController.cs
@@ -1,5 +1,5 @@
 using UnityEngine;
-using Source.Shared;
+using Source.Shared.Services;
 using Source.Shared.Utilities;
 using Source.GamePlay.Services;
 using System;
