+++ Assets/Scripts/GamePlay/Services/GamePlayService.cs
@@ -1,4 +1,4 @@
-using Source.Shared;
+using Source.Shared.Services;
 using Source.Shared.Utilities;
 using System;
 using UnityEngine;
