+++ Assets/Scripts/Shared/Services.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: fe12241d20f0b354e8f03c26e94f3936
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
